+++ R/MsBackendTimsTof.R
@@ -10,13 +10,21 @@
 #' are created with the `MsBackendTimsTof` function. To ensure a small memory
 #' footprint, only general information is kept in memory (such as number of
 #' frames and scans) and all data (specifically the peaks data) is retrieved
-#' from the original file on-the-fly.
+#' from the original file on-the-fly. By extending the [MsBackendCached()]
+#' backend from the `Spectra` package, adding or (locally) changing spectra
+#' values is also supported.
 #'
 #' @section Available methods:
 #'
 #' The following methods are implemented:
 #'
 #' - `$`: access any of the `spectraVariables` of the backend.
+#'
+#' - `$<-`: add a new spectra variable or change values for an existing spectra
+#'   variables. Values can be changed for any spectra variable except *peaks
+#'   variables* ([peaksVariables()]) or special internal variables `"file"` and
+#'   `"frameId"`. Note that changes to spectra variables are only cached within
+#'   the object but not propagated to the original data files.
 #'
 #' - `[`: subset the backend. Only subsetting by element (*row*/`i`) is
 #'   allowed. First the `@indices` slot of `object` is subsetted and then the
@@ -29,11 +37,9 @@
 #'   is supposed to be called right after creating a `MsBackendTimsTof` object
 #'   with `MsBackendTimsTof` function.
 #'
-#' - `dataOrigin`: gets a `character` of length equal to the number of spectra
+#' - `dataStorage`: gets a `character` of length equal to the number of spectra
 #'   in `object` with the names of the '*.d' folders where each spectrum is
 #'   stored.
-#'
-#' - `dataStorage`: same as `dataOrigin`.
 #'
 #' - `intensity`: gets the intensity values from the spectra in the backend.
 #'   Returns a [NumericList()] of `numeric` vectors (intensity values for each
@@ -61,14 +67,28 @@
 #'   vector (length equal to the number of spectra) with the retention time
 #'   for each spectrum.
 #'
+#' - `selectSpectraVariables`: reduces the available spectra variables to the
+#'   ones specified with parameter `spectraVariables`. For *core spectra
+#'   variables* ([coreSpectraVariables()]) only their values will be removed,
+#'   but not the variable itself.
+#'
 #' - `spectraData`: gets spectra variables (specified by `columns`) from
 #'   `object`.
+#'
+#' - `spectraNames`: returns an *ID*/name for each spectrum. As IDs the index
+#'   of the spectrum within the object after the initialization is used. This
+#'   index/spectra name is unique and stable for each spectrum within the same
+#'   object.
 #'
 #' - `spectraVariables`: returns a `character` vector with the spectra variables
 #'   names of core spectra variables defined in the Spectra package and other
 #'   additional variables contained in `object`. Note that also `"mz"` and
 #'   `"intensity"` (which are by default not returned by the
 #'   `spectraVariables,Spectra` method) are returned.
+#'
+#' - `tic`: calculates the total ion count from the intensities of each
+#'   spectrum (for `initial = FALSE`). For `initial = TRUE` `NA` is returned
+#'   for all spectra.
 #'
 #' @param BPPARAM Parameter object defining the parallel processing
 #'   setup to import data in parallel. Defaults to `BPPARAM = bpparam()`.
@@ -82,6 +102,10 @@
 #'
 #' @param files `character` specifying TimsTOF ’*.d’ folders names.
 #'
+#' @param initial For `tic`: `logical(1)` whether the original total ion count
+#'     should be returned (`initial = TRUE`, the default) or whether it
+#'     should be calculated on the spectras' intensities (`initial = FALSE`).
+#'
 #' @param i For `[`: `integer`, `logical` to subset the object.
 #'
 #' @param j For `[`: not supported.
@@ -89,6 +113,11 @@
 #' @param name For `$`: the name of the variable to access.
 #'
 #' @param object `MsBackendTimsTof` object.
+#'
+#' @param spectraVariables `character` with the names of the spectra variables
+#'     that should be retained in the returned object.
+#'
+#' @param value For `$<-`: the value to assign to the spectra variable.
 #'
 #' @param x `MsBackendTimsTof` object.
 #'
@@ -126,20 +155,30 @@
 #' rtime(be_sub)
 #'
 #' pd <- peaksData(be_sub, columns = c("mz", "intensity", "tof", "inv_ion_mobility"))
-setClass("MsBackendTimsTof",
-         contains = "MsBackend",
-         slots = c(frames = "data.frame",
-                   indices = "matrix",
-                   fileNames = "integer"),
-         prototype = prototype(frames = data.frame(),
-                               indices = matrix(nrow = 0, ncol = 3,
-                                                dimnames = list(NULL,
-                                                                c("frame",
-                                                                  "scan",
-                                                                  "file"))),
-                               fileNames = integer(),
-                               readonly = TRUE,
-                               version = "0.1"))
+#'
+#' ## Add a new spectra variable
+#' be$new_var <- seq_along(be)
+#'
+#' head(be$new_var)
+#'
+#' ## Changing values of a spectra variable. Note that these are only changed
+#' ## *locally* within the object, but not in the original data files.
+#' head(rtime(be))
+#' be$rtime <- rtime(be) + 10
+#' head(rtime(be))
+setClass(
+    "MsBackendTimsTof",
+    contains = "MsBackendCached",
+    slots = c(frames = "data.frame",
+              indices = "matrix",
+              fileNames = "integer"),
+    prototype = prototype(
+        frames = data.frame(),
+        indices = matrix(nrow = 0, ncol = 3,
+                         dimnames = list(NULL, c("frame", "scan", "file"))),
+        fileNames = integer(),
+        readonly = TRUE,
+        version = "0.2"))
 
 #' @importFrom methods validObject
 setValidity("MsBackendTimsTof", function(object) {
@@ -153,6 +192,8 @@ setValidity("MsBackendTimsTof", function(object) {
 #' @importFrom BiocParallel bplapply
 #'
 #' @importMethodsFrom Spectra backendInitialize
+#'
+#' @importClassesFrom Spectra MsBackendCached
 #'
 #' @rdname MsBackendTimsTof
 setMethod("backendInitialize", signature = "MsBackendTimsTof",
@@ -168,6 +209,11 @@ setMethod("backendInitialize", signature = "MsBackendTimsTof",
               if (length(msg))
                   stop(msg)
               object <- .initialize(object, files, BPPARAM)
+              object <- callNextMethod(
+                  object, nspectra = nrow(object@indices),
+                  spectraVariables = c(.TIMSTOF_COLUMNS,
+                                       colnames(object@frames),
+                                       "dataOrigin"))
               validObject(object)
               object
           })
@@ -176,7 +222,7 @@ setMethod("backendInitialize", signature = "MsBackendTimsTof",
 #'
 #' @rdname MsBackendTimsTof
 setMethod("length", "MsBackendTimsTof", function(x) {
-    nrow(x@indices)
+    x@nspectra
 })
 
 #' @rdname MsBackendTimsTof
@@ -192,6 +238,8 @@ setMethod(
 setMethod(
     "peaksVariables", "MsBackendTimsTof",
     function(object) {
+        ## TODO: should we return all available, or just the one
+        ## in @spectraVariables
         if (length(object@fileNames)) {
             .list_tims_columns(names(object@fileNames)[1L])
         } else c("mz", "intensity")
@@ -201,19 +249,29 @@ setMethod(
 #'
 #' @rdname MsBackendTimsTof
 setMethod("mz", "MsBackendTimsTof", function(object) {
-    NumericList(.get_tims_columns(object, "mz"), compress = FALSE)
+    if ("mz" %in% object@spectraVariables)
+        NumericList(.get_tims_columns(object, "mz"), compress = FALSE)
+    else spectraData(object, "mz")[, 1L]
 })
 
 #' @importFrom IRanges NumericList
 #'
 #' @rdname MsBackendTimsTof
 setMethod("intensity", "MsBackendTimsTof", function(object) {
-    NumericList(.get_tims_columns(object, "intensity"), compress = FALSE)
+    if ("intensity" %in% object@spectraVariables)
+        NumericList(.get_tims_columns(object, "intensity"), compress = FALSE)
+    else spectraData(object, "intensity")[, 1L]
 })
 
 #' @rdname MsBackendTimsTof
 setMethod("rtime", "MsBackendTimsTof", function(object) {
-    .get_frame_columns(object, "rtime")
+    if ("rtime" %in% colnames(object@localData))
+        object@localData$rtime
+    else {
+        if ("rtime" %in% object@spectraVariables)
+            .get_frame_columns(object, "rtime")
+        else spectraData(object, "rtime")[, 1L]
+    }
 })
 
 #' @importFrom methods "slot<-"
@@ -232,6 +290,7 @@ setMethod("[", "MsBackendTimsTof", function(x, i, j, ..., drop = FALSE) {
                        paste(x@frames$frameId, x@frames$file)), , drop = FALSE]
     slot(x, "fileNames", check = FALSE) <-
         x@fileNames[x@fileNames %in% unique(x@frames$file)]
+    x <- callNextMethod(x, i = i)
     x
 })
 
@@ -240,26 +299,9 @@ setMethod("[", "MsBackendTimsTof", function(x, i, j, ..., drop = FALSE) {
 #' @rdname MsBackendTimsTof
 setMethod("dataStorage", "MsBackendTimsTof", function(object) {
     if("file" %in% colnames(object@indices) && length(object@fileNames))
-        return(names(object@fileNames[match(object@indices[, "file"],
-                                            object@fileNames)]))
-    character(0)
-})
-
-#' @importMethodsFrom Spectra dataOrigin
-#'
-#' @rdname MsBackendTimsTof
-setMethod("dataOrigin", "MsBackendTimsTof", function(object) {
-    dataStorage(object)
-})
-
-#' @importMethodsFrom Spectra spectraVariables
-#'
-#' @importFrom Spectra coreSpectraVariables
-#'
-#' @rdname MsBackendTimsTof
-setMethod("spectraVariables", "MsBackendTimsTof", function(object) {
-    unique(c(names(coreSpectraVariables()), .TIMSTOF_COLUMNS,
-             colnames(object@frames)))
+        names(object@fileNames[match(object@indices[, "file"],
+                                     object@fileNames)])
+    else character(0)
 })
 
 #' @importMethodsFrom Spectra spectraData
@@ -294,15 +336,78 @@ setMethod("show", "MsBackendTimsTof", function(object) {
 #'
 #' @rdname MsBackendTimsTof
 setMethod("msLevel", "MsBackendTimsTof", function(object, ...) {
-    .get_msLevel(object)
+    if ("msLevel" %in% object@spectraVariables)
+        .get_msLevel(object)
+    else spectraData(object, "msLevel")[, 1L]
 })
 
 #' @rdname MsBackendTimsTof
 setMethod("$", "MsBackendTimsTof", function(x, name) {
-    if (!any(spectraVariables(x) == name))
+    if (!name %in% spectraVariables(x))
         stop("spectra variable '", name, "' not available")
-    if (name == "inv_ion_mobility")
+    if (name == "inv_ion_mobility" &&
+        "inv_ion_mobility" %in% x@spectraVariables)
         .inv_ion_mobility(x)
     else
         spectraData(x, name)[, 1L]
+})
+
+#' @importMethodsFrom Spectra spectraVariables
+#'
+#' @rdname MsBackendTimsTof
+setMethod("spectraVariables", "MsBackendTimsTof", function(object, ...) {
+    union(callNextMethod(), .TIMSTOF_COLUMNS)
+})
+
+#' @rdname MsBackendTimsTof
+setMethod(
+    "selectSpectraVariables", "MsBackendTimsTof",
+    function(object, spectraVariables = spectraVariables(object)) {
+        req_cols <- c("frameId", "file")
+        if ("msLevel" %in% spectraVariables) req_cols <- c(req_cols, "MsMsType")
+        keep <- colnames(object@frames) %in% union(spectraVariables, req_cols)
+        object@frames <- object@frames[, keep, drop = FALSE]
+        callNextMethod(object, spectraVariables)
+    })
+
+#' @rdname MsBackendTimsTof
+#'
+#' @export
+setReplaceMethod("$", "MsBackendTimsTof", function(x, name, value) {
+    if (name %in% union(peaksVariables(x), c("file", "frameId")))
+        stop("Replacing spectra variable \"", name, "\" is not supported.")
+    callNextMethod()
+})
+
+#' @rdname MsBackendTimsTof
+#'
+#' @importMethodsFrom ProtGenerics precScanNum
+#'
+#' @exportMethod precScanNum
+setMethod("precScanNum", "MsBackendTimsTof", function(object) {
+    spectraData(object, "precScanNum")[, 1L]
+})
+
+#' @exportMethod spectraNames
+#'
+#' @importMethodsFrom ProtGenerics spectraNames
+#'
+#' @rdname MsBackendTimsTof
+setMethod("spectraNames", "MsBackendTimsTof", function(object) {
+    row.names(object@indices)
+})
+
+#' @importMethodsFrom ProtGenerics tic
+#'
+#' @importFrom Spectra intensity
+#'
+#' @importFrom MsCoreUtils vapply1d
+#'
+#' @exportMethod tic
+#'
+#' @rdname MsBackendTimsTof
+setMethod("tic", "MsBackendTimsTof", function(object, initial = TRUE) {
+    if (initial)
+        rep(NA_real_, length(object))
+    else vapply1d(intensity(object), sum, na.rm = TRUE)
 })